Moved components into HOME component. Auth redirects now point to login.
