+++ app/src/auth.js
@@ -2,7 +2,7 @@ export default function auth($transitions) {
   $transitions.onStart({to: state => !state.data === 'public'}, trans => {
     const userService = trans.injector().get('userService');
     if (!userService.isAuthenticated()) {
-      return trans.router.stateService.target('welcome');
+      return trans.router.stateService.target('login');
     }
   });
 }
