+++ app/src/components/home/friends/friends.html
