+++ app/src/components/home/friends/friends.js
