+++ app/src/components/home/home.html
@@ -1,0 +1,7 @@
+<app-header current="$ctrl.current"></app-header>
+<nav>
+  <a ui-sref="home.collection">Collection</a>
+  <a ui-sref="home.friends">Friends</a>
+  <a ui-sref="home.gamenights">Gamenights</a>
+</nav>
+<ui-view></ui-view>
