+++ app/src/components/home/home.js
@@ -1,4 +1,4 @@
-import template from './users.html';
+import template from './home.html';
 
 export default {
   template,
