--- app/src/components/users/users.html
@@ -1,7 +1,0 @@
-<app-header current="$ctrl.current"></app-header>
-<nav>
-  <a ui-sref="users.collection">Collection</a>
-  <a ui-sref="users.friends">Friends</a>
-  <a ui-sref="users.gamenights">Gamenights</a>
-</nav>
-<ui-view></ui-view>
